fix: 친구 추가시 무한 친구 요청시 또는 이미 친구인데 또 친구 요청을 할 경우 에러메세지를 띄우도록 만들었습니다.

+++ src/main/java/com/fiveman/newsfeed/friend/service/FriendService.java
@@ -4,14 +4,11 @@ import com.fiveman.newsfeed.common.entity.Friend;
 import com.fiveman.newsfeed.common.entity.User;
 import com.fiveman.newsfeed.friend.dto.FriendListResponseDto;
 import com.fiveman.newsfeed.friend.dto.FriendResponseDto;
-import com.fiveman.newsfeed.friend.dto.FriendRequestDto;
 import com.fiveman.newsfeed.friend.repository.FriendRepository;
-import com.fiveman.newsfeed.user.repository.UserRepository;
 import com.fiveman.newsfeed.user.service.UserService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -28,6 +25,17 @@ public class FriendService {
 
     @Transactional
     public void create(Long fromUserId, Long toUserId) {
+
+        isValidSelf(fromUserId, toUserId);
+
+        List<Friend> friendPending = friendRepository.findByStatusIsPendingAndFromUserId(fromUserId,toUserId);
+        List<Friend> friendAccept = friendRepository.findByStatusIsAcceptedAndFromUserId(fromUserId,toUserId);
+
+        if(!friendAccept.isEmpty()) {
+            throw new RuntimeException("이미 친구입니다.");
+        } else if (!friendPending.isEmpty()) {
+            throw new RuntimeException("이미 보내진 요청입니다");
+        }
 
         User fromUser = userService.findById(fromUserId);
         User toUser = userService.findById(toUserId);
@@ -54,7 +62,14 @@ public class FriendService {
 
     public void deleteFriend(Long fromUserId, Long toUserId) {
 
+        isValidSelf(fromUserId, toUserId);
         isValid(fromUserId, toUserId);
+
+        List<Friend> friendAccept = friendRepository.findByStatusIsAcceptedAndFromUserId(fromUserId,toUserId);
+
+        if(friendAccept.isEmpty()) {
+            throw new RuntimeException("잘못된 요청입니다.");
+        }
 
         friendRepository.deleteFriendByFromUserToUser(fromUserId,toUserId);
         friendRepository.deleteFriendByFromUserToUser(toUserId,fromUserId);
@@ -64,6 +79,7 @@ public class FriendService {
     @Transactional
     public void acceptFriendRequest(Long fromUserId, Long toUserId) {
 
+        isValidSelf(fromUserId, toUserId);
         isValid(fromUserId, toUserId);
 
         User fromUser = userService.findById(fromUserId);
@@ -78,18 +94,22 @@ public class FriendService {
     @Transactional
     public void deleteFriendRequest(Long fromUserId, Long toUserId) {
 
+        isValidSelf(fromUserId, toUserId);
         isValid(fromUserId, toUserId);
 
         friendRepository.deleteFriendByFromUserToUser(fromUserId, toUserId);
     }
 
+    public void isValidSelf(Long fromUserId, Long toUserId) {
+
+        if(Objects.equals(fromUserId, toUserId)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "자기 자신에게 요청을 보낼 수 없습니다.");
+        }
+    }
+
     public void isValid(Long fromUserId, Long toUserId) {
         if(!friendRepository.existsByFromUser_UserIdAndToUser_UserId(fromUserId, toUserId)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "찾을 수 없는 요청입니다.");
-        }
-
-        else if(Objects.equals(fromUserId, toUserId)) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "자기 자신에게 요청을 보낼 수 없습니다.");
         }
     }
 }
